Create README.md
# End-to-end-Machine-Learning-Project--Telecom-churn-customer
WE HAVE THE DETAILS OF CUSTOMERS THOSE WHO HAVE CHURN AND NOT CHURN CUSTOMERS. THE TOTAL SIZE OF THE CUSTOMERS IS AROUND THE 20K AND 2 YEARS OF DATA USING MANCHINE LEARNING ALGORITHMS BY USING LOGISTIC REGRESSION(BIARY CLASSIFICATIONS) TO FIND THE PROBABILITY OF CUSTOMERS CHURN AND ALSO THE ACCURANY,PRESCION AND RECALL. ETC
Telecom Churn Case Study
With 21 predictor variables we need to predict whether a particular customer will switch to another telecom provider or not. In telecom terminology, this is referred to as churning and not churning, respectively.